Restructured containers accordingly
